KAFKA-14365: Extract common logic from Fetcher into FetcherUtils
Moving externally stateful entities to be parameters

Removing ConsumerNetworkClient, ConsumerMetadata, and SubscriptionState as Fetcher state and passing them in as parameters where needed.

Introduced FetchConfig

Introduced FetchState and FetchUtils
