+++ clients/src/main/java/org/apache/kafka/clients/consumer/internals/CompletedFetch.java
@@ -1,0 +1,362 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.kafka.clients.consumer.internals;
+
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.kafka.common.IsolationLevel;
+import org.apache.kafka.common.KafkaException;
+import org.apache.kafka.common.TopicPartition;
+import org.apache.kafka.common.errors.CorruptRecordException;
+import org.apache.kafka.common.errors.RecordDeserializationException;
+import org.apache.kafka.common.errors.SerializationException;
+import org.apache.kafka.common.header.Headers;
+import org.apache.kafka.common.header.internals.RecordHeaders;
+import org.apache.kafka.common.message.FetchResponseData;
+import org.apache.kafka.common.record.ControlRecordType;
+import org.apache.kafka.common.record.Record;
+import org.apache.kafka.common.record.RecordBatch;
+import org.apache.kafka.common.record.TimestampType;
+import org.apache.kafka.common.serialization.Deserializer;
+import org.apache.kafka.common.utils.BufferSupplier;
+import org.apache.kafka.common.utils.CloseableIterator;
+import org.apache.kafka.common.utils.LogContext;
+import org.apache.kafka.common.utils.Utils;
+import org.slf4j.Logger;
+
+import java.nio.ByteBuffer;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Optional;
+import java.util.PriorityQueue;
+import java.util.Set;
+
+public class CompletedFetch<K, V> {
+
+    private final Logger log;
+    private final FetchConfig<K, V> fetchConfig;
+    private final SubscriptionState subscriptions;
+    private final TopicPartition partition;
+    private final Iterator<? extends RecordBatch> batches;
+    private final Set<Long> abortedProducerIds;
+    private final PriorityQueue<FetchResponseData.AbortedTransaction> abortedTransactions;
+    private final FetchResponseData.PartitionData partitionData;
+    private final FetchResponseMetricAggregator metricAggregator;
+    private final short responseVersion;
+    private final BufferSupplier decompressionBufferSupplier;
+
+    private int recordsRead;
+    private int bytesRead;
+    private RecordBatch currentBatch;
+    private Record lastRecord;
+    private CloseableIterator<Record> records;
+    private long nextFetchOffset;
+    private Optional<Integer> lastEpoch;
+    private boolean isConsumed = false;
+    private Exception cachedRecordException = null;
+    private boolean corruptLastRecord = false;
+    private boolean initialized = false;
+
+    CompletedFetch(LogContext logContext,
+                   FetchConfig<K, V> fetchConfig,
+                   SubscriptionState subscriptions,
+                   TopicPartition partition,
+                   FetchResponseData.PartitionData partitionData,
+                   FetchResponseMetricAggregator metricAggregator,
+                   Iterator<? extends RecordBatch> batches,
+                   Long fetchOffset,
+                   short responseVersion,
+                   BufferSupplier decompressionBufferSupplier) {
+        this.log = logContext.logger(CompletedFetch.class);
+        this.fetchConfig = fetchConfig;
+        this.subscriptions = subscriptions;
+        this.partition = partition;
+        this.partitionData = partitionData;
+        this.metricAggregator = metricAggregator;
+        this.batches = batches;
+        this.nextFetchOffset = fetchOffset;
+        this.decompressionBufferSupplier = decompressionBufferSupplier;
+        this.responseVersion = responseVersion;
+        this.lastEpoch = Optional.empty();
+        this.abortedProducerIds = new HashSet<>();
+        this.abortedTransactions = abortedTransactions(partitionData);
+    }
+
+    TopicPartition partition() {
+        return partition;
+    }
+
+    FetchResponseData.PartitionData partitionData() {
+        return partitionData;
+    }
+
+    FetchResponseMetricAggregator metricAggregator() {
+        return metricAggregator;
+    }
+
+    short responseVersion() {
+        return responseVersion;
+    }
+
+    long nextFetchOffset() {
+        return nextFetchOffset;
+    }
+
+    boolean isConsumed() {
+        return isConsumed;
+    }
+
+    Optional<Integer> lastEpoch() {
+        return lastEpoch;
+    }
+
+    boolean initialized() {
+        return this.initialized;
+    }
+
+    void setInitialized() {
+        this.initialized = true;
+    }
+
+    void drain() {
+        if (!isConsumed) {
+            maybeCloseRecordStream();
+            cachedRecordException = null;
+            this.isConsumed = true;
+            this.metricAggregator.record(partition, bytesRead, recordsRead);
+
+            // we move the partition to the end if we received some bytes. This way, it's more likely that partitions
+            // for the same topic can remain together (allowing for more efficient serialization).
+            if (bytesRead > 0)
+                subscriptions.movePartitionToEnd(partition);
+        }
+    }
+
+    private void maybeEnsureValid(RecordBatch batch) {
+        if (fetchConfig.checkCrcs() && currentBatch.magic() >= RecordBatch.MAGIC_VALUE_V2) {
+            try {
+                batch.ensureValid();
+            } catch (CorruptRecordException e) {
+                throw new KafkaException("Record batch for partition " + partition + " at offset " +
+                        batch.baseOffset() + " is invalid, cause: " + e.getMessage());
+            }
+        }
+    }
+
+    private void maybeEnsureValid(Record record) {
+        if (fetchConfig.checkCrcs()) {
+            try {
+                record.ensureValid();
+            } catch (CorruptRecordException e) {
+                throw new KafkaException("Record for partition " + partition + " at offset " + record.offset()
+                        + " is invalid, cause: " + e.getMessage());
+            }
+        }
+    }
+
+    private void maybeCloseRecordStream() {
+        if (records != null) {
+            records.close();
+            records = null;
+        }
+    }
+
+    private Record nextFetchedRecord() {
+        while (true) {
+            if (records == null || !records.hasNext()) {
+                maybeCloseRecordStream();
+
+                if (!batches.hasNext()) {
+                    // Message format v2 preserves the last offset in a batch even if the last record is removed
+                    // through compaction. By using the next offset computed from the last offset in the batch,
+                    // we ensure that the offset of the next fetch will point to the next batch, which avoids
+                    // unnecessary re-fetching of the same batch (in the worst case, the consumer could get stuck
+                    // fetching the same batch repeatedly).
+                    if (currentBatch != null)
+                        nextFetchOffset = currentBatch.nextOffset();
+                    drain();
+                    return null;
+                }
+
+                currentBatch = batches.next();
+                lastEpoch = currentBatch.partitionLeaderEpoch() == RecordBatch.NO_PARTITION_LEADER_EPOCH ?
+                        Optional.empty() : Optional.of(currentBatch.partitionLeaderEpoch());
+
+                maybeEnsureValid(currentBatch);
+
+                if (fetchConfig.isolationLevel() == IsolationLevel.READ_COMMITTED && currentBatch.hasProducerId()) {
+                    // remove from the aborted transaction queue all aborted transactions which have begun
+                    // before the current batch's last offset and add the associated producerIds to the
+                    // aborted producer set
+                    consumeAbortedTransactionsUpTo(currentBatch.lastOffset());
+
+                    long producerId = currentBatch.producerId();
+                    if (containsAbortMarker(currentBatch)) {
+                        abortedProducerIds.remove(producerId);
+                    } else if (isBatchAborted(currentBatch)) {
+                        log.debug("Skipping aborted record batch from partition {} with producerId {} and " +
+                                        "offsets {} to {}",
+                                partition, producerId, currentBatch.baseOffset(), currentBatch.lastOffset());
+                        nextFetchOffset = currentBatch.nextOffset();
+                        continue;
+                    }
+                }
+
+                records = currentBatch.streamingIterator(decompressionBufferSupplier);
+            } else {
+                Record record = records.next();
+                // skip any records out of range
+                if (record.offset() >= nextFetchOffset) {
+                    // we only do validation when the message should not be skipped.
+                    maybeEnsureValid(record);
+
+                    // control records are not returned to the user
+                    if (!currentBatch.isControlBatch()) {
+                        return record;
+                    } else {
+                        // Increment the next fetch offset when we skip a control batch.
+                        nextFetchOffset = record.offset() + 1;
+                    }
+                }
+            }
+        }
+    }
+
+    List<ConsumerRecord<K, V>> fetchRecords(int maxRecords) {
+        // Error when fetching the next record before deserialization.
+        if (corruptLastRecord)
+            throw new KafkaException("Received exception when fetching the next record from " + partition
+                    + ". If needed, please seek past the record to "
+                    + "continue consumption.", cachedRecordException);
+
+        if (isConsumed)
+            return Collections.emptyList();
+
+        List<ConsumerRecord<K, V>> records = new ArrayList<>();
+        try {
+            for (int i = 0; i < maxRecords; i++) {
+                // Only move to next record if there was no exception in the last fetch. Otherwise we should
+                // use the last record to do deserialization again.
+                if (cachedRecordException == null) {
+                    corruptLastRecord = true;
+                    lastRecord = nextFetchedRecord();
+                    corruptLastRecord = false;
+                }
+                if (lastRecord == null)
+                    break;
+                records.add(parseRecord(partition, currentBatch, lastRecord));
+                recordsRead++;
+                bytesRead += lastRecord.sizeInBytes();
+                nextFetchOffset = lastRecord.offset() + 1;
+                // In some cases, the deserialization may have thrown an exception and the retry may succeed,
+                // we allow user to move forward in this case.
+                cachedRecordException = null;
+            }
+        } catch (SerializationException se) {
+            cachedRecordException = se;
+            if (records.isEmpty())
+                throw se;
+        } catch (KafkaException e) {
+            cachedRecordException = e;
+            if (records.isEmpty())
+                throw new KafkaException("Received exception when fetching the next record from " + partition
+                        + ". If needed, please seek past the record to "
+                        + "continue consumption.", e);
+        }
+        return records;
+    }
+
+    private void consumeAbortedTransactionsUpTo(long offset) {
+        if (abortedTransactions == null)
+            return;
+
+        while (!abortedTransactions.isEmpty() && abortedTransactions.peek().firstOffset() <= offset) {
+            FetchResponseData.AbortedTransaction abortedTransaction = abortedTransactions.poll();
+
+            if (abortedTransaction != null)
+                abortedProducerIds.add(abortedTransaction.producerId());
+        }
+    }
+
+    private boolean isBatchAborted(RecordBatch batch) {
+        return batch.isTransactional() && abortedProducerIds.contains(batch.producerId());
+    }
+
+    private PriorityQueue<FetchResponseData.AbortedTransaction> abortedTransactions(FetchResponseData.PartitionData partition) {
+        if (partition.abortedTransactions() == null || partition.abortedTransactions().isEmpty())
+            return null;
+
+        PriorityQueue<FetchResponseData.AbortedTransaction> abortedTransactions = new PriorityQueue<>(
+                partition.abortedTransactions().size(), Comparator.comparingLong(FetchResponseData.AbortedTransaction::firstOffset)
+        );
+        abortedTransactions.addAll(partition.abortedTransactions());
+        return abortedTransactions;
+    }
+
+    private boolean containsAbortMarker(RecordBatch batch) {
+        if (!batch.isControlBatch())
+            return false;
+
+        Iterator<Record> batchIterator = batch.iterator();
+        if (!batchIterator.hasNext())
+            return false;
+
+        Record firstRecord = batchIterator.next();
+        return ControlRecordType.ABORT == ControlRecordType.parse(firstRecord.key());
+    }
+
+    /**
+     * Parse the record entry, deserializing the key / value fields if necessary
+     */
+    private ConsumerRecord<K, V> parseRecord(TopicPartition partition,
+                                             RecordBatch batch,
+                                             Record record) {
+        Deserializer<K> keyDeserializer = fetchConfig.keyDeserializer();
+        Deserializer<V> valueDeserializer = fetchConfig.valueDeserializer();
+
+        try {
+            long offset = record.offset();
+            long timestamp = record.timestamp();
+            Optional<Integer> leaderEpoch = maybeLeaderEpoch(batch.partitionLeaderEpoch());
+            TimestampType timestampType = batch.timestampType();
+            Headers headers = new RecordHeaders(record.headers());
+            ByteBuffer keyBytes = record.key();
+            byte[] keyByteArray = keyBytes == null ? null : org.apache.kafka.common.utils.Utils.toArray(keyBytes);
+            K key = keyBytes == null ? null : keyDeserializer.deserialize(partition.topic(), headers, keyByteArray);
+            ByteBuffer valueBytes = record.value();
+            byte[] valueByteArray = valueBytes == null ? null : Utils.toArray(valueBytes);
+            V value = valueBytes == null ? null : valueDeserializer.deserialize(partition.topic(), headers, valueByteArray);
+            return new ConsumerRecord<>(partition.topic(), partition.partition(), offset,
+                    timestamp, timestampType,
+                    keyByteArray == null ? ConsumerRecord.NULL_SIZE : keyByteArray.length,
+                    valueByteArray == null ? ConsumerRecord.NULL_SIZE : valueByteArray.length,
+                    key, value, headers, leaderEpoch);
+        } catch (RuntimeException e) {
+            throw new RecordDeserializationException(partition, record.offset(),
+                    "Error deserializing key/value for partition " + partition +
+                            " at offset " + record.offset() + ". If needed, please seek past the record to continue consumption.", e);
+        }
+    }
+
+    private Optional<Integer> maybeLeaderEpoch(int leaderEpoch) {
+        return leaderEpoch == RecordBatch.NO_PARTITION_LEADER_EPOCH ? Optional.empty() : Optional.of(leaderEpoch);
+    }
+
+}
